% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fndat-package.R
\docType{package}
\name{fndat-package}
\alias{fndat}
\alias{fndat-package}
\title{fndat: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/david-beauchesne/fndat}
  \item Report bugs at \url{https://github.com/david-beauchesne/fndat/issues}
}

}
\author{
\strong{Maintainer}: David Beauchesne \email{david.beauchesne@hotmail.com} (\href{https://orcid.org/0000-0002-3590-8161}{ORCID}) [copyright holder]

}
\keyword{internal}
