% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keysearch.R
\name{keysearch}
\alias{keysearch}
\title{Function to perform and update resarch for resources on open government base on certain keywords}
\usage{
keysearch()
}
\value{
A list of resources
}
\description{
This function is used to research resources on the canadian open government portal
(https://open.canada.ca/en) using the \code{rgovcan} package. It allows the user to perform a
using a series of keywords and store the results. The results can then be marked as verified
so that the next research will discard them automatically and either show only unverified
resources or new resources available on the portal since the last research.
}
\details{
For now, verified resources must be added manually in './inst/extdata/verified.csv'.
}
